Add framework for common piedpiper storage tools
First we define a base client in storeman which defines the
common functions which must be exposed by all storage backends.
Then we create a minio_client which implements these functions.

We also expose a storage_client function which will instantiate
a storage_client based on the type passed in as a parameter. Currently
the only storage_client type supported is "minio"