Fix wait_for_thread_id_complete
This function will now parse out the 'x-gman-tasks-pending' status
when calling HEAD on the thread to get the status of the thread.
Previously, GMan would return 0 tasks running when an executor FaaS
had not yet picked up the task for execution, effectively putting
the task in a "pending" state and not completed.